Prince/Trustpilot debug (#6383)
